feat: allow plugin to filter allowed polyfills
The default implementation will silently polyfill from the complete list of supported modules.

Often the need for a polyfill may be due to a misconfigured project, where a ‘browser’ import was not properly resolved.

This feature adds support to allow the rollup-project to explitly allow specific polyfills, or simply log their use as an advisory to the developer.

If a polyfill is rejected, the project will expose retain the import as an external reference.